+++ sw.js
@@ -1,4 +1,4 @@
-const CACHE_NAME = "ihnatiev-site-v0.6.19c12-r1";
+const CACHE_NAME = "ihnatiev-site-v0.6.19c13-r1";
 const APP_SHELL = [
   "./",
   "./index.html",
@@ -14,6 +14,7 @@ const APP_SHELL = [
   "./js/gallery-renderer.js",
   "./js/document-lightbox.js",
   "./js/downloads-renderer.js",
+  "./js/content-loader.js",
   "./js/youtube-feed.js",
   "./js/mobile-navigation.js",
   "./js/contact-page.js",